make project importable
