+++ cmd/process-exporter/main.go
@@ -11,9 +11,9 @@ import (
 	"time"
 
 	"github.com/ncabatoff/fakescraper"
-	common "github.com/ncabatoff/process-exporter"
+	"github.com/ncabatoff/process-exporter/collector"
+	"github.com/ncabatoff/process-exporter/common"
 	"github.com/ncabatoff/process-exporter/config"
-	"github.com/ncabatoff/process-exporter/proc"
 	"github.com/prometheus/client_golang/prometheus"
 )
 
@@ -65,152 +65,6 @@ Config file process selection (filename.yml):
 ` + "\n")
 
 }
-
-var (
-	numprocsDesc = prometheus.NewDesc(
-		"namedprocess_namegroup_num_procs",
-		"number of processes in this group",
-		[]string{"groupname"},
-		nil)
-
-	cpuUserSecsDesc = prometheus.NewDesc(
-		"namedprocess_namegroup_cpu_user_seconds_total",
-		"Cpu user usage in seconds",
-		[]string{"groupname"},
-		nil)
-
-	cpuSystemSecsDesc = prometheus.NewDesc(
-		"namedprocess_namegroup_cpu_system_seconds_total",
-		"Cpu system usage in seconds",
-		[]string{"groupname"},
-		nil)
-
-	readBytesDesc = prometheus.NewDesc(
-		"namedprocess_namegroup_read_bytes_total",
-		"number of bytes read by this group",
-		[]string{"groupname"},
-		nil)
-
-	writeBytesDesc = prometheus.NewDesc(
-		"namedprocess_namegroup_write_bytes_total",
-		"number of bytes written by this group",
-		[]string{"groupname"},
-		nil)
-
-	majorPageFaultsDesc = prometheus.NewDesc(
-		"namedprocess_namegroup_major_page_faults_total",
-		"Major page faults",
-		[]string{"groupname"},
-		nil)
-
-	minorPageFaultsDesc = prometheus.NewDesc(
-		"namedprocess_namegroup_minor_page_faults_total",
-		"Minor page faults",
-		[]string{"groupname"},
-		nil)
-
-	contextSwitchesDesc = prometheus.NewDesc(
-		"namedprocess_namegroup_context_switches_total",
-		"Context switches",
-		[]string{"groupname", "ctxswitchtype"},
-		nil)
-
-	membytesDesc = prometheus.NewDesc(
-		"namedprocess_namegroup_memory_bytes",
-		"number of bytes of memory in use",
-		[]string{"groupname", "memtype"},
-		nil)
-
-	openFDsDesc = prometheus.NewDesc(
-		"namedprocess_namegroup_open_filedesc",
-		"number of open file descriptors for this group",
-		[]string{"groupname"},
-		nil)
-
-	worstFDRatioDesc = prometheus.NewDesc(
-		"namedprocess_namegroup_worst_fd_ratio",
-		"the worst (closest to 1) ratio between open fds and max fds among all procs in this group",
-		[]string{"groupname"},
-		nil)
-
-	startTimeDesc = prometheus.NewDesc(
-		"namedprocess_namegroup_oldest_start_time_seconds",
-		"start time in seconds since 1970/01/01 of oldest process in group",
-		[]string{"groupname"},
-		nil)
-
-	numThreadsDesc = prometheus.NewDesc(
-		"namedprocess_namegroup_num_threads",
-		"Number of threads",
-		[]string{"groupname"},
-		nil)
-
-	statesDesc = prometheus.NewDesc(
-		"namedprocess_namegroup_states",
-		"Number of processes in states Running, Sleeping, Waiting, Zombie, or Other",
-		[]string{"groupname", "state"},
-		nil)
-
-	scrapeErrorsDesc = prometheus.NewDesc(
-		"namedprocess_scrape_errors",
-		"general scrape errors: no proc metrics collected during a cycle",
-		nil,
-		nil)
-
-	scrapeProcReadErrorsDesc = prometheus.NewDesc(
-		"namedprocess_scrape_procread_errors",
-		"incremented each time a proc's metrics collection fails",
-		nil,
-		nil)
-
-	scrapePartialErrorsDesc = prometheus.NewDesc(
-		"namedprocess_scrape_partial_errors",
-		"incremented each time a tracked proc's metrics collection fails partially, e.g. unreadable I/O stats",
-		nil,
-		nil)
-
-	threadWchanDesc = prometheus.NewDesc(
-		"namedprocess_namegroup_threads_wchan",
-		"Number of threads in this group waiting on each wchan",
-		[]string{"groupname", "wchan"},
-		nil)
-
-	threadCountDesc = prometheus.NewDesc(
-		"namedprocess_namegroup_thread_count",
-		"Number of threads in this group with same threadname",
-		[]string{"groupname", "threadname"},
-		nil)
-
-	threadCpuSecsDesc = prometheus.NewDesc(
-		"namedprocess_namegroup_thread_cpu_seconds_total",
-		"Cpu user/system usage in seconds",
-		[]string{"groupname", "threadname", "cpumode"},
-		nil)
-
-	threadIoBytesDesc = prometheus.NewDesc(
-		"namedprocess_namegroup_thread_io_bytes_total",
-		"number of bytes read/written by these threads",
-		[]string{"groupname", "threadname", "iomode"},
-		nil)
-
-	threadMajorPageFaultsDesc = prometheus.NewDesc(
-		"namedprocess_namegroup_thread_major_page_faults_total",
-		"Major page faults for these threads",
-		[]string{"groupname", "threadname"},
-		nil)
-
-	threadMinorPageFaultsDesc = prometheus.NewDesc(
-		"namedprocess_namegroup_thread_minor_page_faults_total",
-		"Minor page faults for these threads",
-		[]string{"groupname", "threadname"},
-		nil)
-
-	threadContextSwitchesDesc = prometheus.NewDesc(
-		"namedprocess_namegroup_thread_context_switches_total",
-		"Context switches for these threads",
-		[]string{"groupname", "threadname", "ctxswitchtype"},
-		nil)
-)
 
 type (
 	prefixRegex struct {
@@ -348,7 +202,7 @@ func main() {
 		matchnamer = namemapper
 	}
 
-	pc, err := NewProcessCollector(*procfsPath, *children, matchnamer, *recheck, *debug)
+	pc, err := collector.NewProcessCollector(*procfsPath, *children, matchnamer, *recheck, *debug)
 	if err != nil {
 		log.Fatalf("Error initializing: %v", err)
 	}
@@ -380,192 +234,4 @@ func main() {
 	if err := http.ListenAndServe(*listenAddress, nil); err != nil {
 		log.Fatalf("Unable to setup HTTP server: %v", err)
 	}
-}
-
-type (
-	scrapeRequest struct {
-		results chan<- prometheus.Metric
-		done    chan struct{}
-	}
-
-	NamedProcessCollector struct {
-		scrapeChan chan scrapeRequest
-		*proc.Grouper
-		source               proc.Source
-		scrapeErrors         int
-		scrapeProcReadErrors int
-		scrapePartialErrors  int
-		debug                bool
-	}
-)
-
-func NewProcessCollector(
-	procfsPath string,
-	children bool,
-	n common.MatchNamer,
-	recheck bool,
-	debug bool,
-) (*NamedProcessCollector, error) {
-	fs, err := proc.NewFS(procfsPath, debug)
-	if err != nil {
-		return nil, err
-	}
-	p := &NamedProcessCollector{
-		scrapeChan: make(chan scrapeRequest),
-		Grouper:    proc.NewGrouper(n, children, recheck, debug),
-		source:     fs,
-		debug:      debug,
-	}
-
-	colErrs, _, err := p.Update(p.source.AllProcs())
-	if err != nil {
-		if debug {
-			log.Print(err)
-		}
-		return nil, err
-	}
-	p.scrapePartialErrors += colErrs.Partial
-	p.scrapeProcReadErrors += colErrs.Read
-
-	go p.start()
-
-	return p, nil
-}
-
-// Describe implements prometheus.Collector.
-func (p *NamedProcessCollector) Describe(ch chan<- *prometheus.Desc) {
-	ch <- cpuUserSecsDesc
-	ch <- cpuSystemSecsDesc
-	ch <- numprocsDesc
-	ch <- readBytesDesc
-	ch <- writeBytesDesc
-	ch <- membytesDesc
-	ch <- openFDsDesc
-	ch <- worstFDRatioDesc
-	ch <- startTimeDesc
-	ch <- majorPageFaultsDesc
-	ch <- minorPageFaultsDesc
-	ch <- contextSwitchesDesc
-	ch <- numThreadsDesc
-	ch <- statesDesc
-	ch <- scrapeErrorsDesc
-	ch <- scrapeProcReadErrorsDesc
-	ch <- scrapePartialErrorsDesc
-	ch <- threadWchanDesc
-	ch <- threadCountDesc
-	ch <- threadCpuSecsDesc
-	ch <- threadIoBytesDesc
-	ch <- threadMajorPageFaultsDesc
-	ch <- threadMinorPageFaultsDesc
-	ch <- threadContextSwitchesDesc
-}
-
-// Collect implements prometheus.Collector.
-func (p *NamedProcessCollector) Collect(ch chan<- prometheus.Metric) {
-	req := scrapeRequest{results: ch, done: make(chan struct{})}
-	p.scrapeChan <- req
-	<-req.done
-}
-
-func (p *NamedProcessCollector) start() {
-	for req := range p.scrapeChan {
-		ch := req.results
-		p.scrape(ch)
-		req.done <- struct{}{}
-	}
-}
-
-func (p *NamedProcessCollector) scrape(ch chan<- prometheus.Metric) {
-	permErrs, groups, err := p.Update(p.source.AllProcs())
-	p.scrapePartialErrors += permErrs.Partial
-	if err != nil {
-		p.scrapeErrors++
-		log.Printf("error reading procs: %v", err)
-	} else {
-		for gname, gcounts := range groups {
-			ch <- prometheus.MustNewConstMetric(numprocsDesc,
-				prometheus.GaugeValue, float64(gcounts.Procs), gname)
-			ch <- prometheus.MustNewConstMetric(membytesDesc,
-				prometheus.GaugeValue, float64(gcounts.Memory.ResidentBytes), gname, "resident")
-			ch <- prometheus.MustNewConstMetric(membytesDesc,
-				prometheus.GaugeValue, float64(gcounts.Memory.VirtualBytes), gname, "virtual")
-			ch <- prometheus.MustNewConstMetric(membytesDesc,
-				prometheus.GaugeValue, float64(gcounts.Memory.VmSwapBytes), gname, "swapped")
-			ch <- prometheus.MustNewConstMetric(startTimeDesc,
-				prometheus.GaugeValue, float64(gcounts.OldestStartTime.Unix()), gname)
-			ch <- prometheus.MustNewConstMetric(openFDsDesc,
-				prometheus.GaugeValue, float64(gcounts.OpenFDs), gname)
-			ch <- prometheus.MustNewConstMetric(worstFDRatioDesc,
-				prometheus.GaugeValue, float64(gcounts.WorstFDratio), gname)
-			ch <- prometheus.MustNewConstMetric(cpuUserSecsDesc,
-				prometheus.CounterValue, gcounts.CPUUserTime, gname)
-			ch <- prometheus.MustNewConstMetric(cpuSystemSecsDesc,
-				prometheus.CounterValue, gcounts.CPUSystemTime, gname)
-			ch <- prometheus.MustNewConstMetric(readBytesDesc,
-				prometheus.CounterValue, float64(gcounts.ReadBytes), gname)
-			ch <- prometheus.MustNewConstMetric(writeBytesDesc,
-				prometheus.CounterValue, float64(gcounts.WriteBytes), gname)
-			ch <- prometheus.MustNewConstMetric(majorPageFaultsDesc,
-				prometheus.CounterValue, float64(gcounts.MajorPageFaults), gname)
-			ch <- prometheus.MustNewConstMetric(minorPageFaultsDesc,
-				prometheus.CounterValue, float64(gcounts.MinorPageFaults), gname)
-			ch <- prometheus.MustNewConstMetric(contextSwitchesDesc,
-				prometheus.CounterValue, float64(gcounts.CtxSwitchVoluntary), gname, "voluntary")
-			ch <- prometheus.MustNewConstMetric(contextSwitchesDesc,
-				prometheus.CounterValue, float64(gcounts.CtxSwitchNonvoluntary), gname, "nonvoluntary")
-			ch <- prometheus.MustNewConstMetric(numThreadsDesc,
-				prometheus.GaugeValue, float64(gcounts.NumThreads), gname)
-			ch <- prometheus.MustNewConstMetric(statesDesc,
-				prometheus.GaugeValue, float64(gcounts.States.Running), gname, "Running")
-			ch <- prometheus.MustNewConstMetric(statesDesc,
-				prometheus.GaugeValue, float64(gcounts.States.Sleeping), gname, "Sleeping")
-			ch <- prometheus.MustNewConstMetric(statesDesc,
-				prometheus.GaugeValue, float64(gcounts.States.Waiting), gname, "Waiting")
-			ch <- prometheus.MustNewConstMetric(statesDesc,
-				prometheus.GaugeValue, float64(gcounts.States.Zombie), gname, "Zombie")
-			ch <- prometheus.MustNewConstMetric(statesDesc,
-				prometheus.GaugeValue, float64(gcounts.States.Other), gname, "Other")
-
-			for wchan, count := range gcounts.Wchans {
-				ch <- prometheus.MustNewConstMetric(threadWchanDesc,
-					prometheus.GaugeValue, float64(count), gname, wchan)
-			}
-
-			for _, thr := range gcounts.Threads {
-				ch <- prometheus.MustNewConstMetric(threadCountDesc,
-					prometheus.GaugeValue, float64(thr.NumThreads),
-					gname, thr.Name)
-				ch <- prometheus.MustNewConstMetric(threadCpuSecsDesc,
-					prometheus.CounterValue, float64(thr.CPUUserTime),
-					gname, thr.Name, "user")
-				ch <- prometheus.MustNewConstMetric(threadCpuSecsDesc,
-					prometheus.CounterValue, float64(thr.CPUSystemTime),
-					gname, thr.Name, "system")
-				ch <- prometheus.MustNewConstMetric(threadIoBytesDesc,
-					prometheus.CounterValue, float64(thr.ReadBytes),
-					gname, thr.Name, "read")
-				ch <- prometheus.MustNewConstMetric(threadIoBytesDesc,
-					prometheus.CounterValue, float64(thr.WriteBytes),
-					gname, thr.Name, "write")
-				ch <- prometheus.MustNewConstMetric(threadMajorPageFaultsDesc,
-					prometheus.CounterValue, float64(thr.MajorPageFaults),
-					gname, thr.Name)
-				ch <- prometheus.MustNewConstMetric(threadMinorPageFaultsDesc,
-					prometheus.CounterValue, float64(thr.MinorPageFaults),
-					gname, thr.Name)
-				ch <- prometheus.MustNewConstMetric(threadContextSwitchesDesc,
-					prometheus.CounterValue, float64(thr.CtxSwitchVoluntary),
-					gname, thr.Name, "voluntary")
-				ch <- prometheus.MustNewConstMetric(threadContextSwitchesDesc,
-					prometheus.CounterValue, float64(thr.CtxSwitchNonvoluntary),
-					gname, thr.Name, "nonvoluntary")
-			}
-		}
-	}
-	ch <- prometheus.MustNewConstMetric(scrapeErrorsDesc,
-		prometheus.CounterValue, float64(p.scrapeErrors))
-	ch <- prometheus.MustNewConstMetric(scrapeProcReadErrorsDesc,
-		prometheus.CounterValue, float64(p.scrapeProcReadErrors))
-	ch <- prometheus.MustNewConstMetric(scrapePartialErrorsDesc,
-		prometheus.CounterValue, float64(p.scrapePartialErrors))
 }
